Rj Added

+++ client/src/Page/Admin/Invoice.jsx
@@ -1,65 +1,66 @@
-// Invoice.js
-
-import React, { useEffect, useState } from 'react'
-import { Link } from 'react-router-dom'
-import InvoiceFormat from '../../Component/InvoiceFormat'
-import { RiDeleteBin6Line } from 'react-icons/ri'
-
-import search from '/search.svg'
-import { BsThreeDotsVertical } from 'react-icons/bs'
-
-import { FaRegEye, FaCheck } from 'react-icons/fa6'
-import { FaPlus } from 'react-icons/fa6'
-import ManualInovoice from '../../Component/ManualInovoice'
-import SearchBar from '../../Component/SearchBar'
-import { useDispatch, useSelector } from 'react-redux'
+import React, { useEffect, useState } from 'react';
+import { Link } from 'react-router-dom';
+import InvoiceFormat from '../../Component/InvoiceFormat';
+import { RiDeleteBin6Line } from 'react-icons/ri';
+import search from '/search.svg';
+import { BsThreeDotsVertical } from 'react-icons/bs';
+import { FaRegEye, FaCheck } from 'react-icons/fa6';
+import { FaPlus } from 'react-icons/fa6';
+import ManualInovoice from '../../Component/ManualInovoice';
+import SearchBar from '../../Component/SearchBar';
+import { useDispatch, useSelector } from 'react-redux';
 import {
   deleteInvoices,
   editInvoices,
   fetchInvoices,
   searchInvoice,
-} from '../../features/invoice'
+} from '../../features/invoice';
 
 const Invoice = () => {
-  const [update, setUpdate] = useState(false)
-  const [filter, setFilter] = useState('')
-  const [status, setStatus] = useState(false)
-  const user = useSelector((state) => state.auth.user)
-  const msg = useSelector((state) => state.invoice.msg)
-  const [modal, setModal] = useState(false)
-  const [dropdownIndex, setDropdownIndex] = useState(null) // To keep track of which dropdown is open
-  const dispatch = useDispatch()
-  const invoices = useSelector((state) => state.invoice.data)
+  const [update, setUpdate] = useState(false);
+  const [filter, setFilter] = useState('');
+  const [status, setStatus] = useState(false);
+  const user = useSelector((state) => state.auth.user);
+  const msg = useSelector((state) => state.invoice.msg);
+  const [modal, setModal] = useState(false);
+  const [dropdownIndex, setDropdownIndex] = useState(null); // To keep track of which dropdown is open for each item
+  const dispatch = useDispatch();
+  const invoices = useSelector((state) => state.invoice.data);
+
   const handleDelete = async (id) => {
-    dispatch(deleteInvoices(id))
-    setUpdate(true)
-  }
+    const confirmed = window.confirm('Are you sure you want to delete this item?');
+    if (confirmed) {
+      dispatch(deleteInvoices(id));
+      setUpdate(true);
+    }
+  };
 
   const handleSearch = (e) => {
-    setFilter(e.target.value)
-  }
+    setFilter(e.target.value);
+  };
+
   const handlePaid = (id) => {
-    dispatch(editInvoices(id, true))
-    setUpdate(true)
-  }
-  console.log(status)
+    dispatch(editInvoices(id, true));
+    setUpdate(true);
+  };
+
   useEffect(() => {
     if (filter && filter !== '') {
-      dispatch(searchInvoice(filter))
+      dispatch(searchInvoice(filter));
     } else {
-      dispatch(fetchInvoices())
-      setUpdate(false)
+      dispatch(fetchInvoices());
+      setUpdate(false);
     }
-  }, [filter, update, msg, status])
+  }, [filter, update, msg, status]);
 
-  const handleDropdownClick = (index) => {
-    setDropdownIndex(index === dropdownIndex ? null : index)
-  }
+  const handleDropdownClick = (id) => {
+    setDropdownIndex(id === dropdownIndex ? null : id);
+  };
 
   return (
     <>
       <div className="w-full h-full bg-white1">
-        <div className=" w-11/12 m-auto h-full flex flex-col gap-2 ">
+        <div className=" w-11/12 m-auto h-full flex flex-col gap-2  ">
           <h1 className="font-bold pt-5 tracking-wider">DOCUMENTS / INVOICE</h1>
           <div className="w-full h-full max-h-20 flex flex-col md:flex-row gap-2 py-5 items-center justify-between ">
             <div className="w-full md:w-fit">
@@ -85,9 +86,9 @@ const Invoice = () => {
               </button>
             </div>
           </div>
-          <div className="w-full h-full ">
+          <div className="w-full h-80 ">
             <div className="relative w-full h-full max-h-[660px] bg-white overflow-y-scroll border-primary-color border-l-4 border-b-4">
-              <table className="w-full h-auto  ">
+              <table className="w-full h-fit  overflow-y-auto  ">
                 <thead className="sticky top-0 bg-primary-color z-10">
                   <tr className="text-center text-xs md:text-base">
                     <th className="text-white p-3 w-1/5">Reference Number</th>
@@ -101,7 +102,7 @@ const Invoice = () => {
                 <tbody className="w-full h-full bg-white font-regular">
                   {user && user.role === 'Admin'
                     ? invoices
-                        ?.filter((item) => item.status == status)
+                        ?.filter((item) => item.status === status)
                         .map((val, index) => (
                           <tr
                             key={index}
@@ -120,7 +121,11 @@ const Invoice = () => {
                             </td>
                             <td className="p-2">{(val?.amount).toFixed(2)}</td>
                             <th
-                              className={`font-semibold ${val?.status === false ? 'text-red' : 'text-primary-color'} p-2 w-1/5`}
+                              className={`font-semibold ${
+                                val?.status === false
+                                  ? 'text-red'
+                                  : 'text-primary-color'
+                              } p-2 w-1/5`}
                             >
                               {val?.status === false ? 'Unpaid' : 'Paid'}
                             </th>
@@ -128,17 +133,19 @@ const Invoice = () => {
                               <div className="relative">
                                 <button
                                   className="relative focus:outline-none my-3"
-                                  onClick={() => handleDropdownClick(index)}
+                                  onClick={() => handleDropdownClick(val?._id)}
                                 >
                                   <BsThreeDotsVertical size={25} />
                                 </button>
-                                {dropdownIndex === index && (
-                                  <ul className="absolute w-36 right-0 top-7 bg-white shadow-md rounded-md z-10">
+                                {dropdownIndex === val?._id && (
+                                  <ul className={`absolute w-36 right-0 bg-white shadow-md rounded-md z-10 `}>
                                     {val?.status === false && (
                                       <>
                                         <li>
                                           <button
-                                            onClick={() => handlePaid(val?._id)}
+                                            onClick={() =>
+                                              handlePaid(val?._id)
+                                            }
                                             className=" px-4 py-2 text-sm text-primary-color flex items-center gap-3 hover:bg-gray w-full"
                                           >
                                             <FaCheck /> Paid
@@ -146,9 +153,7 @@ const Invoice = () => {
                                         </li>
                                         <li>
                                           <button
-                                            onClick={() =>
-                                              handleDelete(val?._id)
-                                            }
+                                            onClick={() => handleDelete(val?._id)}
                                             className=" px-4 py-2 text-sm text-primary-color flex items-center gap-3  hover:bg-gray w-full"
                                           >
                                             <RiDeleteBin6Line /> Delete
@@ -173,7 +178,7 @@ const Invoice = () => {
                     : invoices
                         ?.filter(
                           (item) =>
-                            item.status == status &&
+                            item.status === status &&
                             item.tenant_id.user_id._id === user.id,
                         )
                         .map((val, index) => (
@@ -194,7 +199,11 @@ const Invoice = () => {
                             </td>
                             <td className="p-2">{(val?.amount).toFixed(2)}</td>
                             <th
-                              className={`font-semibold ${val?.status === false ? 'text-red' : 'text-primary-color'} p-3 w-1/5`}
+                              className={`font-semibold ${
+                                val?.status === false
+                                  ? 'text-red'
+                                  : 'text-primary-color'
+                              } p-3 w-1/5`}
                             >
                               {val?.status === false ? 'Unpaid' : 'Paid'}
                             </th>
@@ -202,17 +211,19 @@ const Invoice = () => {
                               <div className="relative">
                                 <button
                                   className="relative focus:outline-none my-3"
-                                  onClick={() => handleDropdownClick(index)}
+                                  onClick={() => handleDropdownClick(val?._id)}
                                 >
                                   <BsThreeDotsVertical size={25} />
                                 </button>
-                                {dropdownIndex === index && (
-                                  <ul className="absolute w-36 right-0 top-7 bg-white shadow-md rounded-md z-10">
+                                {dropdownIndex === val?._id && (
+                                  <ul className={`absolute w-36 right-0 bg-white shadow-md rounded-md z-10 ${index === invoices.length - 1 ? 'bottom-12' : (index === invoices.length - 2 ? 'bottom-24' : 'top-12')}`}>
                                     {val?.status === false && (
                                       <>
                                         <li>
                                           <button
-                                            onClick={() => handlePaid(val?._id)}
+                                            onClick={() =>
+                                              handlePaid(val?._id)
+                                            }
                                             className=" px-4 py-2 text-sm text-primary-color flex items-center gap-3 hover:bg-gray w-full"
                                           >
                                             <FaCheck /> Paid
@@ -220,9 +231,7 @@ const Invoice = () => {
                                         </li>
                                         <li>
                                           <button
-                                            onClick={() =>
-                                              handleDelete(val?._id)
-                                            }
+                                            onClick={() => handleDelete(val?._id)}
                                             className=" px-4 py-2 text-sm text-primary-color flex items-center gap-3  hover:bg-gray w-full"
                                           >
                                             <RiDeleteBin6Line /> Delete
@@ -252,7 +261,7 @@ const Invoice = () => {
       </div>
       {modal && <ManualInovoice setModal={setModal} />}
     </>
-  )
-}
+  );
+};
 
-export default Invoice
+export default Invoice;
